Separate the Duke and DukeException classes
